neotest-gtest just does not work
I was unable to make it work with automatic building.
Neotest saw changes in test only after second execution, which was
annoying.
Moreover, there is a race condition somewhere in neotest during
initialization of neovim that makes it randonly not working.
So in order to have working neotest, I had to restart neovim a few
times.
Maybe I will examine it more closely, but at the moment it's just a
waste of time.